requirements.txt
flask
python-dotenv
requests
google-generativeai
Pillow 